+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove user-data feature.
 - Move user-data to a standard rust implementation inspired by 0.13 serenity.
 - Merged `serenity-native` and `serenity-rustls` into a single `serenity` feature.
+- Implement `PlayerContext::play()`
 
 ## 0.10.0